Three the compiler found that nothing else could
NavigationPath.append is generic over Hashable & Codable, so a bare  has
no context to infer from -- the compiler settled on  and reported that
it has no member , which is a true statement about the wrong type. Named
the route.

 multiplies a CGFloat by a Double and both have
overloads that could apply, so it is ambiguous rather than wrong. Converted
explicitly. This is in ArchShape, which draws the arch on the roster screen -- the
oldest geometry in the project.

A memberwise initialiser is positional however many labels it carries, so the
injected stores had to be declared where they are passed. They sit next to
 now, which is where they belonged anyway: they are the same kind of
thing.

+++ Arch/Components/ArchShape.swift
@@ -6,8 +6,10 @@ import SwiftUI
 /// 180 degrees is the left springing point, 270 the crown, 360 the right springing point.
 private func pointOnCircle(_ centre: CGPoint, _ radius: CGFloat, _ angle: Angle) -> CGPoint {
     CGPoint(
-        x: centre.x + radius * cos(angle.radians),
-        y: centre.y + radius * sin(angle.radians)
+        // `cos` returns a Double and `radius` is a CGFloat, and the multiply has
+        // overloads for both -- ambiguous rather than wrong. Converted explicitly.
+        x: centre.x + radius * CGFloat(cos(angle.radians)),
+        y: centre.y + radius * CGFloat(sin(angle.radians))
     )
 }
